+++ CHANGELOG.md
@@ -5,7 +5,9 @@ If a change is prefixed with one or more of these, it has special mean:
 
 # master
 
-- added documentation to model's `reset()` method [#29]
+- added options parameter to collection's `toJSON()` method that is passed when calling `toJSON()` on the models in the collection [#28]
+- added options parameter to model's `toJSON()` method and now you can define the properties you want converted [#28]
+- added documentation for model's `reset()` method [#29]
 - [breaking change] a bunch of API have been changes, see documentation for what the APIs are
 - transaction support
 - restructure of unit/integration tests